Refactor SettingsChannel
