+++ shell/platform/tizen/channels/encodable_value_holder.h
@@ -2,6 +2,9 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
+#ifndef EMBEDDER_ENCODABLE_VALUE_HOLDER_H_
+#define EMBEDDER_ENCODABLE_VALUE_HOLDER_H_
+
 #include "flutter/shell/platform/common/client_wrapper/include/flutter/encodable_value.h"
 
 namespace flutter {
@@ -28,3 +31,5 @@ struct EncodableValueHolder {
 };
 
 }  // namespace flutter
+
+#endif  // EMBEDDER_ENCODABLE_VALUE_HOLDER_H_
